- added libdrawtext-noft project to the visual studio solution - added examples/nofreetype project to the visual studio solution

+++ src/utf8.c
@@ -117,7 +117,7 @@ size_t dtx_utf8_from_char_code(int code, char *buf)
 	}
 
 	if(!nbytes && buf) {
-		for(i=0; i<nbytes; i++) {
+		for(i=0; i<(int)nbytes; i++) {
 			int idx = nbytes - i - 1;
 			int mask, shift, prefix;
